Update code.py
A working version of the integration. The date and time variables have been moved into the "State" class to serve as global variables. The log files show time and date as 0 because the variables are not being updated in the child classes. Need to implement the allowance of child classes to update the variable in the parent class. It appears that the easiest path is using a "Global" declaration in "State".