Update code.py
Rev_4, each state is writing a little bit of relevant data to the SD card. Need to determine if the time difference will be handled in the machine or in Excel. In the machine will require more research into creating global variables that store and can manipulate the data. These key concepts are the "Scope" of the variables, the property function using "get", and "set", also "global", "local" and "nonlocal".